Python3.11 on FreeBSD throws UnboundLocalError
cannot access local variable 'cur_row_idx' where it is not associated with
a value.

Python3.10 on GNU/Linux does not throw :-/

Setting to -1 in the event the read does not occur, to avoid returning a
non zero value. After the read, number of lines calculation has a plus 1
to account for the 0 based counter. start=1 could be used in enumerate
after adjusting the way the writing happens in write_DomainPointers(). it
currently works with 0 based indexing.